+++ src/gfx/gauge.h
@@ -6,8 +6,8 @@ public:
     enum DIRECTION {GAUGE_UP, GAUGE_DOWN, GAUGE_RIGHT, GAUGE_LEFT, GAUGE_TIME};
 private:  enum DIRECTION dir;
 public: Gauge( const char *file, DIRECTION up );
-    void Draw( float percentage );
-    void SetSize( float x, float y );
-    void GetSize( float &x, float &y );
+    void Draw( double percentage );
+    void SetSize( double x, double y );
+    void GetSize( double &x, double &y );
 };
