+++ src/gui/glut_support.cpp
@@ -167,9 +167,9 @@ void DrawGlutMouse( int mousex, int mousey, VSSprite *spr )
     GFXEnable( TEXTURE0 );
     GFXDisable( DEPTHTEST );
     GFXDisable( LIGHTING );
-    float sizex = 0, sizey = 0;
+    double sizex = 0, sizey = 0;
     spr->GetSize( sizex, sizey );
-    float tempx = 0, tempy = 0;
+    double tempx = 0, tempy = 0;
     spr->GetPosition( tempx, tempy );
     spr->SetPosition( tempx+ -1+.5*sizex+float(mousex)
                      /(.5*g_game.x_resolution), tempy+1+.5*sizey-float(mousey)/(.5*g_game.y_resolution) );
